[1.1:14.03]: added decomposition to smaller components
