+++ src/App.css
@@ -2,3 +2,13 @@
     margin: 20px auto 20px auto;
     max-width: 400px;
 }
+.app-add-form {
+    padding: 25px;
+    background-color: #020204;
+    border-radius: 4px;
+    box-shadow: 15px 15px 30px rgba(0, 0, 0, 0.15);
+    color: #22b455;
+}
+.app-add-form .form-header {
+    text-align: center;
+}
