--- src/components/candidate-add-form/candidate-add-form.css
@@ -1,86 +1,0 @@
-.app-add-form {
-  padding: 25px;
-  background-color: #020204;
-  border-radius: 4px;
-  box-shadow: 15px 15px 30px rgba(0, 0, 0, 0.15);
-  color: #22b455;
-}
-.app-add-form .form-header {
-  text-align: center;
-}
-.app-add-form .form {
-  display: flex;
-  flex-direction: column;
-}
-.app-add-form .form label {
-  display: block;
-  margin-bottom: 10px;
-  font-weight: bold;
-}
-.app-add-form .form input,
-.app-add-form .form textarea {
-  display: block;
-  padding: 10px;
-  margin-bottom: 20px;
-  border-radius: 5px;
-  border: 1px solid #ccc;
-  resize: none;
-}
-.app-add-form .form input:focus,
-.app-add-form .form textarea:focus {
-  outline: none;
-  border-color: #22b455;
-  box-shadow: 0 0 5px #22b455;
-}
-.app-add-form .form input[type=date],
-.app-add-form .form textarea[type=date] {
-  color: #ffffff;
-}
-.app-add-form .form input[type=date]::before,
-.app-add-form .form textarea[type=date]::before {
-  content: attr(placeholder);
-  position: absolute;
-  color: #777474;
-}
-.app-add-form .form input[type=date]:focus, .app-add-form .form input[type=date] :valid,
-.app-add-form .form textarea[type=date]:focus,
-.app-add-form .form textarea[type=date] :valid {
-  color: #666666;
-}
-.app-add-form .form input[type=date]:focus::before, .app-add-form .form input[type=date] :valid::before,
-.app-add-form .form textarea[type=date]:focus::before,
-.app-add-form .form textarea[type=date] :valid::before {
-  content: "";
-}
-.app-add-form .form .btn-group {
-  display: flex;
-  justify-content: flex-end;
-  column-gap: 10px;
-}
-.app-add-form .form .btn-group button {
-  padding: 10px 20px;
-  border-radius: 5px;
-  border: none;
-  color: #fff;
-  font-size: 16px;
-  font-weight: bold;
-  text-transform: uppercase;
-  cursor: pointer;
-  transition: background-color 0.2s ease-in-out;
-}
-.app-add-form .form .btn-group button.save {
-  background-color: #22b455;
-}
-.app-add-form .form .btn-group button.save:hover {
-  background-color: #0051ff;
-}
-.app-add-form .form .btn-group button.discard {
-  background-color: #dc3545;
-}
-.app-add-form .form .btn-group button.discard:hover {
-  color: #dc3545;
-  background-color: #fff;
-  outline: 1px solid #dc3545;
-}
-
-/*# sourceMappingURL=candidate-add-form.css.map */
