--- src/components/candidate-add-form/candidate-add-form.scss
@@ -1,90 +1,0 @@
-.app-add-form {
-  padding: 25px;
-  background-color: #020204;
-  border-radius: 4px;
-  box-shadow: 15px 15px 30px rgba(0,0,0, .15);
-  color: #22b455;
-  .form-header {
-    text-align: center;
-  }
-  .form {
-    display: flex;
-    flex-direction: column;
-
-    label {
-      display: block;
-      margin-bottom: 10px;
-      font-weight: bold;
-    }
-
-    input,
-    textarea {
-      display: block;
-      padding: 10px;
-      margin-bottom: 20px;
-      border-radius: 5px;
-      border: 1px solid #ccc;
-      resize: none;
-
-      &:focus {
-        outline: none;
-        border-color: #22b455;
-        box-shadow: 0 0 5px #22b455;
-      }
-
-      &[type="date"] {
-        color: #ffffff;
-
-        &::before {
-          content: attr(placeholder);
-          position: absolute;
-          color: #777474;
-        }
-
-        &:focus, :valid {
-          color: #666666;
-
-          &::before {
-            content: "";
-          }
-        }
-      }
-
-    }
-
-    .btn-group {
-      display: flex;
-      justify-content: flex-end;
-      column-gap: 10px;
-      button {
-        padding: 10px 20px;
-        border-radius: 5px;
-        border: none;
-        color: #fff;
-        font-size: 16px;
-        font-weight: bold;
-        text-transform: uppercase;
-        cursor: pointer;
-        transition: background-color 0.2s ease-in-out;
-
-        &.save {
-          background-color: #22b455;
-
-          &:hover {
-            background-color: #0051ff;
-          }
-        }
-
-        &.discard {
-          background-color: #dc3545;
-
-          &:hover {
-            color: #dc3545;
-            background-color: #fff;
-            outline: 1px solid #dc3545;
-          }
-        }
-      }
-    }
-  }
-}
