+++ src/components/form-button-list/FormButtonList.Module.css
@@ -1,0 +1,25 @@
+.app-add-form .form .btn-group button {
+    padding: 10px 20px;
+    border-radius: 5px;
+    border: none;
+    color: #fff;
+    font-size: 16px;
+    font-weight: bold;
+    text-transform: uppercase;
+    cursor: pointer;
+    transition: background-color 0.2s ease-in-out;
+}
+.app-add-form .form .btn-group button.save {
+    background-color: #22b455;
+}
+.app-add-form .form .btn-group button.save:hover {
+    background-color: #0051ff;
+}
+.app-add-form .form .btn-group button.discard {
+    background-color: #dc3545;
+}
+.app-add-form .form .btn-group button.discard:hover {
+    color: #dc3545;
+    background-color: #fff;
+    outline: 1px solid #dc3545;
+}
