+++ src/components/form-textarea-list/FormTextareaList.Module.css
@@ -1,0 +1,18 @@
+.app-add-form .form label {
+    display: block;
+    margin-bottom: 10px;
+    font-weight: bold;
+}
+.app-add-form .form textarea {
+    display: block;
+    padding: 10px;
+    margin-bottom: 20px;
+    border-radius: 5px;
+    border: 1px solid #ccc;
+    resize: none;
+}
+.app-add-form .form textarea:focus {
+    outline: none;
+    border-color: #22b455;
+    box-shadow: 0 0 5px #22b455;
+}
