+++ src/components/main-form/MainForm.Module.css
@@ -1,0 +1,9 @@
+.app-add-form .form {
+    display: flex;
+    flex-direction: column;
+}
+.app-add-form .form .btn-group {
+    display: flex;
+    justify-content: flex-end;
+    column-gap: 10px;
+}
